+++ CHANGELOG.md
@@ -3,6 +3,13 @@ Shazarr notable changes.
 
 [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) format.
 
+## 📦 0.0.9
+### 🚀 Added
+* Playwright testing
+* Github CI Eslint + testing
+### 🐛 Fixed
+* Fix history order issue on delete / recognize offline record
+
 ## 📦 0.0.8
 ### 🐛 Fixed
 * Fix missing "not found" message